Move drawer to fungi
