+++ apps/parsley/.env-cmdrc.json
@@ -72,7 +72,7 @@
     "REACT_APP_GRAPHQL_URL": "https://evergreen.mongodb.com/graphql/query",
     "REACT_APP_LOGKEEPER_URL": "https://logkeeper2.build.10gen.cc",
     "REACT_APP_PARSLEY_URL": "https://parsley-local.corp.mongodb.com",
-    "REACT_APP_SPRUCE_URL": "https://spruce.mongodb.com"
-
+    "REACT_APP_SPRUCE_URL": "https://spruce.mongodb.com",
+    "VITE_PARSLEY_AI_URL": "https://sage.staging.corp.mongodb.com"
   }
 }
