+++ apps/parsley/package.json
@@ -52,7 +52,6 @@
     "@leafygreen-ui/checkbox": "^13.1.2",
     "@leafygreen-ui/confirmation-modal": "^5.0.8",
     "@leafygreen-ui/copyable": "^8.1.0",
-    "@leafygreen-ui/drawer": "^4.2.1",
     "@leafygreen-ui/emotion": "^4.0.7",
     "@leafygreen-ui/empty-state": "^3.0.4",
     "@leafygreen-ui/guide-cue": "^8.0.4",
