+++ apps/parsley/src/components/Chatbot/index.tsx
@@ -1,12 +1,34 @@
-import { ChatbotContextProvider } from "./Context";
-import { ChatDrawer } from "./Drawer";
+import { ChatDrawer } from "@evg-ui/fungi/ChatDrawer";
+import { ChatFeed } from "@evg-ui/fungi/ChatFeed";
+import { ChatProvider } from "@evg-ui/fungi/ChatProvider";
+import { useLogContext } from "context/LogContext";
+import { parsleyChatURL } from "utils/environmentVariables";
 
 interface Props {
   children: React.ReactNode;
 }
 
-export const Chatbot: React.FC<Props> = ({ children }) => (
-  <ChatbotContextProvider>
-    <ChatDrawer>{children}</ChatDrawer>
-  </ChatbotContextProvider>
-);
+export const Chatbot: React.FC<Props> = ({ children }) => {
+  const { logMetadata } = useLogContext();
+  const { execution, logType, origin, taskID } = logMetadata ?? {};
+
+  const bodyData = {
+    logMetadata: {
+      execution: Number(execution),
+      log_type: logType,
+      origin,
+      task_id: taskID,
+    },
+  };
+
+  return (
+    <ChatProvider>
+      <ChatDrawer
+        chatContent={<ChatFeed apiUrl={parsleyChatURL} bodyData={bodyData} />}
+        title="Parsley AI"
+      >
+        {children}
+      </ChatDrawer>
+    </ChatProvider>
+  );
+};
