+++ packages/fungi/package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@ai-sdk/react": "^2.0.11",
+    "@emotion/styled": "11.14.1",
+    "@leafygreen-ui/drawer": "^4.2.1",
     "@lg-chat/chat-window": "^4.1.1",
     "@lg-chat/input-bar": "^10.0.1",
     "@lg-chat/leafygreen-chat-provider": "^5.0.0",
